Totals
Totals(previous balance, expected..) have a new agroupation and the
'Current balance' has be changed to 'Total balance'.

All totalizers repect the filters now. With this when we use the --tag=
filter the 'expenses' only sum transactions with this tag.

The 'total balance' has ben fixed to respect max date filter, because
on select previous months he always showed the current value.